Fixes admin list error when the referenced object no longer exist.
When a `content_object` a comment is referring to no longer exists the
`comment.content_object` call sometimes returns `None` which means a
`get_absolute_url` call will fail. In other occurences if it no longer
exists it can raise an `AttributeError` when trying to retrieve the
object as `get_object_for_this_type` gets called and
`self.model_class()` returns `None` in the line.